Image processing operations done. Very good accuracy with difference comparation
The computation in parallel of the average, difference, ... is
completed. It is remaining the covariance which is not synchronized, I
should use OpenCV to compute it even though it is less efficient.

The results are promising, I achieved a 97% of accuracy (in the paper
with the same dataset they achieved a 100% of accuracy) which is very
good.

Some cleaning to the code will be appreciated, just take 1-2h to clean
it and it should be fine.